+++ src/pages/blog.js
@@ -11,7 +11,12 @@ export default ({ data }) => (
       <Nav />
       <div id={Style.posts}>
         {data.allMarkdownRemark.edges.map(({ node }, index) => (
-          <BlogItem key={index} data={node} index={index}></BlogItem>
+          <BlogItem
+            images={data.images.edges}
+            key={index}
+            data={node}
+            index={index}
+          ></BlogItem>
         ))}
       </div>
     </Layout>
@@ -30,6 +35,20 @@ export const query = graphql`
             date
             thumbnail
           }
+        }
+      }
+    }
+    images: allS3Image {
+      edges {
+        node {
+          file: localFile {
+            image: childImageSharp {
+              fluid(quality: 100) {
+                ...GatsbyImageSharpFluid_withWebp
+              }
+            }
+          }
+          Key
         }
       }
     }